TimelineWriter broken into TimelineWriter and SvgTimelineWriter
